requirements.txt
python-dotenv~=1.0.1
openai~=1.35.13
requests~=2.32.3
pandas~=2.2.2
supabase~=2.6.0
config~=0.5.1
Flask~=3.0.3
PyJWT~=2.8.0